Worked on EF support

+++ src/Abp/Modules/Core/Abp.Modules.Core.Infrastructure.NHibernate/Modules/Core/Data/Migrations/V20140323/_20140323_05_CreatePermissionsTable.cs
@@ -9,7 +9,7 @@ namespace Abp.Modules.Core.Data.Migrations.V20140323
         public override void Up()
         {
             Create.Table("AbpPermissions")
-                .WithColumn("Id").AsInt32().NotNullable().PrimaryKey().Identity()
+                .WithColumn("Id").AsInt64().NotNullable().PrimaryKey().Identity()
                 .WithColumn("RoleId").AsInt32().Nullable().ForeignKey("AbpRoles", "Id")
                 .WithColumn("UserId").AsInt32().Nullable().ForeignKey("AbpUsers", "Id")
                 .WithColumn("Name").AsAnsiString(128).NotNullable()
